Improved pause feedback; desaturated screen and text.
Displaying text in Flash really sucks.  Most of the lines added for this
were to get the text.  I'm also not pleased that I had to pull in Screen
for that part. If I realised it would be like this, I'd have broken the
pausing into its own file.  But it works.  For now?  SHIP IT!